requirements.txt
fastapi
uvicorn
python-dotenv
databricks-sdk==0.65.0
pydantic
pandas
openpyxl
# Testing dependencies
pytest
pytest-asyncio
pytest-cov
httpx
# Databricks tools for secure SQL execution and job orchestration
-e ../ai-dev-kit/databricks-tools-core